fix: avoid circular dependency in data types
